getCard to int instead of long

+++ src/main/java/study/sem_4/Models/BankAccount.java
@@ -23,7 +23,7 @@ public class BankAccount {
                 " }";
     }
 
-	public long getCard() {
+	public int getCard() {
 		return this.card;
 	}
